pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "kaspa-tpu"
version = "0.1.0"
description = "Hybrid CPU/TPU miner for Kaspa's kHeavyHash PoW algorithm"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
]

[project.optional-dependencies]
tpu = [
    "jax[tpu]",
]
grpc = [
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
]
dev = [
    "pytest>=7.0.0",
]
all = [
    "kaspa-tpu[tpu,grpc,dev]",
]

[project.scripts]
kaspa-tpu = "kaspa_tpu.main:main"

[tool.setuptools.packages.find]
include = ["kaspa_tpu*"]
